src/requirements.txt
Mezzanine==3.1.5
Cartridge==0.9.4
requests==2.3.0
untangle==1.1.0
gdata==2.0.18
sorl-thumbnail==11.12
MySQL-python==1.2.5
django-compressor==1.4
python-memcached==1.53
django_extensions
setproctitle
gunicorn
django-annoying